[spi] move intermediary resultset from spi to pinot
